clean

+++ src/main/java/com/example/demo/repository/TaskRepository.java
@@ -1,16 +1,16 @@
 package com.example.demo.repository;
 
-import com.example.demo.domain.Task;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.domain.Pageable;
-
 import java.time.LocalDate;
 import java.util.List;
-import org.springframework.stereotype.Repository;
 
-public interface TaskRepository extends JpaRepository<Task, Long> {
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+
+import com.example.demo.domain.Task;
+
+
+public interface TaskRepository extends JpaRepository<Task, Integer> {
     @Query("SELECT t FROM Task t")
     List<Task> getTaskList(Pageable pageable);
